some text file refactoring + general refactoring

+++ src/model/Packages.java
@@ -24,7 +24,7 @@ public class Packages implements HasID {
      * @param dimensions  the dimensions of this package
      * @param cost        the cost associated with this package
      */
-    public Packages(Integer packageID, double weight, String dimensions, Integer cost) {
+    public Packages(Integer packageID, double weight, String dimensions, double cost) {
         this.cost = cost;
         this.packageID = packageID;
         this.weight = weight;
@@ -146,7 +146,7 @@ public class Packages implements HasID {
         //Integer orderID = parts[1].equals("null") ? null : Integer.parseInt(parts[1]); // Handle null for orderID
         double weight = Double.parseDouble(parts[1]);
         String dimensions = parts[2];
-        int cost = Integer.parseInt(parts[3]);
+        double cost = Double.parseDouble(parts[3]);
 
         // Create a new Packages object
         Packages packages = new Packages(packageID, weight, dimensions, cost);
